#1318 refactor schema mapper functional test to introduce base class
