+++ vuu/src/test/java/org/finos/vuu/util/SchemaJavaExample.java
@@ -44,6 +44,8 @@ public class SchemaJavaExample {
                 //.withFieldsMap(columnNameByExternalField)
                 .build();
 
+
+
         //get data from ignite as list of values
         var queryName = "myQuery";
         var igniteStore = new FakeIgniteStore<SchemaJavaTestData>();
